pyproject.toml
[project]
name = "open-source-buddy"
version = "0.1.0"
description = "Open Source Buddy - AI-powered assistant for open source projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "sentence-transformers",
    "openai-agents>=0.0.17",
    "openai>=1.0.0",
    "httpx",
    "PyGithub>=2.3",
    "numpy",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "ruff",
    "black",
    "mypy",
]

[project.scripts]
open-source-buddy = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "our_agents"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 